api-gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["Maks1mka <maksim_goncharov_2006@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.121.0"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
redis = "^7.0.1"
aiohttp = "^3.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
